requirements.txt
psycopg2==2.8.5
python-dotenv==0.14.0
flask==1.1.2
flask-sqlalchemy==2.4.4
gunicorn==20.0.4
py-healthcheck==1.10.1
prometheus_client==0.8.0